[loader] Split VerifyMD5() into its own file.
menu.c:
- Moved #include <ogc/video.h> to menu.h, since it's needed by the
  ReconfigVideo() prototype.
- Removed #include <ogc/video_types.h>, since it isn't actually used.
